Changes to Level and Stage Structure
Edited the LevelLoadCallback prefab to now be the Level prefab, which still contains the LevelLoadCallback script, but also a LevelStats script for the GameManager to access the player's start position and playRuns or cycles to beat the level. Also created new fields for the GameManager, including a reference to LevelStats, and the cycleTotal & cycleCurr(ent). Restructured the code regarding initializing and changing stages to be more concise and readable, and the player respawn now takes them back to the first stage.