Partially Implemented Level Cycling
Made it so the level now cycles, however there are still kinks to work out: need to make the obstacles fully reset when the player dies or complete a cycle (and need to make a distinction between the player death and cycle for deteriorating blocks). It may not even be an issue, however, as we may not use grabbable blocks that can be transferred between rooms in this game.