Created IDeteriorate and frameworked DeterioratingPlatform script
IDeteriorate was created to be an interface in case the platform needs to be deteriorated directly. The DeterioratingPlatform script has been frameworked for the game object to be created. The necessary additions will be the changing of sprites which for the time being just has an int (damageStatus) being a placeholder to track the changes.