IProjectile, Projectile, and Arrow scripts created
Projectile is the parent script for all projectile-based scripts and objects. It contains the needed code for its game object to travel the stage with a kinematic rigidbody. IProjectile was set up so the projectile can be instantiated with a given Vector2 velocity. Arrow was then created as a child class of the Projectile class. The Arrow script has the capability of triggering player death and being destroyed upon hitting an obstacle in its path.