A lotta stuff added/updated including the implementation of buffs
Made it so that a pick up object intersecting with a player would trigger the PlayerController to notify the GameManager that they have collected a pick up w/its corresponding information. The GameManager then dissects the information to apply a buff and track it through the CollectPickup method assigning its fields, activePickUp and pickUpTimer, with the correct values. Once the timer reaches 0 in the Update method, disable the pick up's buff. The CollectPickup method is also where the GameManager tells other scripts to be modified according to the buff: JumpHeightBuff will notify the player, enabling the player to jump a multiple of their jump height, and TimeFreeze interacts with a list of IFreezeable interfaces of all freezeable gameObjects in the level. Instances of the interface are added upon the Start method of a freezeable game object, and removed upon its destruction. Currently in the middle of implementing all IFreezable interfaces to the proper scripts.