Created FreezeableFunctionality class that any potential freezing object's scripts extend and created the Remove method to IFreezeable
FreezeableFunctionality makes the incorporation of freezing objects much easier by extending the class and using the isFrozen field to see how the object must be treated. Additionally, the Remove method in IFreezeable allows the GameManager to destroy and remove all the freezeable objects before it itself deletes from quitting the game, preventing errors.