Done a lotta stuff, check description for a stress-addled summary
Deteriorating block fully works now with different sprites as it takes damage, and fades away just prior to being destroyed.

Created a prefab for the StageExit, and on for RiftExit (being the level end) using different sprites and different aspects of the same code.

Consolidated some Enums onto Enums script.

Currently working on ArrowShooter and Arrow objects/scripts. Currently, the rotation for the arrow isn't working.